refactor(api): split idempotency middleware into focused modules (key/store/memory/layer)
Break the 1224-line god-file `middleware/idempotency.rs` into four
single-responsibility files under `middleware/idempotency/`:

- `key.rs`    — IdempotencyKey newtype + validation + cache-key composition
                + SHA-256 body/identity fingerprinting
- `store.rs`  — IdempotencyStore trait + IdempotencyStoreError +
                StorageBackedIdempotencyStore (PG bridge)
- `memory.rs` — InMemoryIdempotencyStore (moka future::Cache)
- `layer.rs`  — IdempotencyLayer + IdempotencyService + metrics +
                request/response buffering + replay-header logic

`mod.rs` holds IdempotencyConfig, the header constants
(IDEMPOTENCY_KEY_HEADER, IDEMPOTENT_REPLAY_HEADER), shared numeric
constants, and `pub use` re-exports that preserve every previously-public
path byte-for-byte. All inline tests relocated next to what they exercise;
zero tests dropped (236 passed, 1 skipped — unchanged count).
